+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   - Change the message of tweet when the user who sends the tweet is not the obtainer.
   - Add `<meta name="title">`
 - `Special:Achievements/<OTHER_USERNAME>` will show the list of the other user's achieved achievements.
+- Add a contributions tool link that links to Special:Achievements.
 
 ### Others